requirements.txt
torch
torchvision
tensorboardX
tensorboard
easydict
pandas
numpy
scipy
matplotlib
py-cpuinfo
